requirements.txt
Flask==2.0.1
Werkzeug==2.3.8
google-generativeai
python-dotenv